+++ framework/CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   
 - Added a `.execute` method on the AuthZ API to execute `CosmosMsg` types on behalf of a granter.
 - `.sub_accounts` method on `Account` for getting Abstract Client Accounts
+- Publish adapter method of Abstract Client Publisher now returns Adapter object
+- Added a `.account_from` method on the `AbstractClient` for retrieving `Account`s.
 
 ### Changed